Extract Warning model
- Warnings are part of the domain modeling, not the controller.
